pyproject.toml
[project]
name = "diplom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "fastapi[all]>=0.120.0",
    "httpx==0.27.0",
    "asyncpg==0.29.0",
    "alembic==1.14.0",
    "passlib[bcrypt]==1.7.4",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "sqlalchemy>=2.0.44",
]
